+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# ActiveAttr 0.2.2 (unreleased) #
+# ActiveAttr 0.2.2 (November 2, 2011) #
 
 * Fixed all instances of modules' #initialize not invoking its superclass
 * Fixed redefining an attribute appending a new AttributeDefinition